ondemand: deep sleep detection via low-power idle notifier
Implement a deep sleep detection mechanism using a low-power idle
notifier introduced in 9487d11.

A new parameter is provided in the ondemand cpufreq governor and
exported in sysfs:

  /sys/devices/system/cpu/cpufreq/ondemand/deep_sleep

The deep sleep detection is implemented using the following heuristic.

When deep sleep detection is enabled (deep_sleep = 1) if we detect that
the previous low-power idle state was more than 4x the sampling rate, it
means that we're probably going to enter to a deep sleep state or we're
just resuming from a deep sleep.

In both cases we completely ignore the cpu activity (considering it to
the minimum load), because probably there's a big chance to get back to
the deep sleep soon. If it happens, with this heuristic, the cpu will
enter the deep sleep state at the minimum frequency, that is good for
saving battery life.

Signed-off-by: Andrea Righi <andrea@betterlinux.com>